Simplify E2E test structure
