+++ .env.example
@@ -1,7 +1,7 @@
 NODE_ENV=development
 BAS_APP_HOST=0.0.0.0
 BAS_APP_PORT=3000
-BAS_APP_ENABLE_CORS=false
+BAS_APP_CORS_ENABLE=false
 BAS_APP_CORS_ORIGIN=
 BAS_REDIS_ENABLE=true
 BAS_REDIS_HOST=redis
