+++ test/fixtures/responses/accesstoken.response.ts
@@ -1,0 +1,9 @@
+export const accessTokenFromApiResponse = {
+  accessToken: 'sample_access_token_from_mocked_blizzapi',
+  source: 'battlenet',
+};
+
+export const accessTokenFromCacheResponse = {
+  accessToken: 'sample_access_token_from_mocked_redis_cache',
+  source: 'battlenet',
+};
