+++ test/fixtures/responses/index.ts
@@ -1,4 +1,5 @@
-export * from './main.default';
-export * from './placeholder';
-export * from './unauthorized';
-export * from './statusDefaultResponse';
+export * from './main.response';
+export * from './placeholder.response';
+export * from './unauthorized.response';
+export * from './status.properties';
+export * from './accesstoken.response';
