+++ test/fixtures/responses/placeholder.response.ts
@@ -1,0 +1,3 @@
+export const placeholderResponse = {
+  foo: 'bar',
+};
