+++ test/fixtures/responses/status.properties.ts
@@ -1,0 +1,1 @@
+export const statusProperties = ['status', 'uptime', 'timestamp'];
