--- test/fixtures/responses/statusDefaultResponse.ts
@@ -1,7 +1,0 @@
-import { ResponseType } from '../types';
-
-type StatusDefaultResponse = ResponseType<200>;
-
-export const statusDefaultResponse: StatusDefaultResponse = {
-  statusCode: 200,
-};
