+++ test/fixtures/responses/unauthorized.response.ts
@@ -1,0 +1,7 @@
+export const unauthorizedResponse = {
+  statusCode: 401,
+  data: {
+    statusCode: 401,
+    message: 'Unauthorized',
+  },
+};
