--- test/fixtures/test-cases/auth/index.ts
@@ -1,8 +1,0 @@
-import { authEnabled } from './enabled';
-import { authDisabled } from './disabled';
-import { TestCaseConfig } from '../../types';
-
-export const auth: TestCaseConfig = {
-  name: 'Authorization',
-  states: [...[authEnabled], ...[authDisabled]],
-};
