--- test/fixtures/test-cases/battlenet/index.ts
@@ -1,8 +1,0 @@
-import { TestCaseConfig } from '../../types';
-import { battlenetConfigured } from './configured';
-import { battlenetUnonfigured } from './unconfigured';
-
-export const battlenet: TestCaseConfig = {
-  name: 'Battle.net',
-  states: [...[battlenetConfigured], ...[battlenetUnonfigured]],
-};
